+++ lib/castle/podcast.ex
@@ -20,15 +20,12 @@ defmodule Castle.Podcast do
     |> cast(attrs, [:account_id, :title, :subtitle, :image_url, :created_at, :updated_at, :published_at])
   end
 
-  def recent(accounts, limit, page) when is_list(accounts) do
-    offset = (page - 1) * limit
-    Castle.Repo.all(from p in Castle.Podcast, where: p.account_id in ^accounts,
-      limit: ^limit, offset: ^offset, order_by: [asc: :title])
+  def recent_query(accounts) do
+    from p in Castle.Podcast, where: p.account_id in ^accounts
   end
 
-  def total(accounts) when is_list(accounts) do
-    Castle.Repo.one(from p in Castle.Podcast, where: p.account_id in ^accounts,
-      select: count("*"))
+  def total(queryable) do
+    Castle.Repo.one(from r in subquery(queryable), select: count(r.id))
   end
 
   def max_updated_at do
